Don't build STM32 firmware lib or libopenstm32.
The 'summon-arm-toolchain' script should do just that per default,
build an ARM toolchain. The STM32-specific libs are not related to
that and many users will not need or want them.

So, default to off for both, but leave in the option so interested users
can enable either of them.